parse timestamps to usec (network + media)

+++ output.h
@@ -8,7 +8,7 @@ struct pulse {
   pa_stream *stream;
 };
 
-pa_usec_t latency_to_usec(int samplerate, int latency);
+pa_usec_t latency_to_usec(int samplerate, uint64_t latency);
 void output_init(struct pulse *pulse);
 void create_stream(struct pulse *pulse, pa_sample_spec *ss);
 void connect_stream(struct pulse *pulse, const pa_buffer_attr *bufattr);
